pyproject.toml
[tool.poetry]
name = "fastapi-hw3"
version = "0.1.0"
description = ""
authors = ["Impe11e <1mpellee@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
alembic = "^1.14.1"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.38"
pydantic-extra-types = "^2.10.2"
phonenumbers = "^8.13.55"
pydantic = {extras = ["email"], version = "^2.10.6"}
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
fastapi-mail = "^1.4.2"
pydantic-settings = "^2.8.0"
redis = "^5.2.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.42.2"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
